Function risp::read

pub fn read<'a>(text: &'a str) -> Result<Vec<RispValue>, Error<Rule>>

Reader

The function read implements a reader, converting a string into an s-expression, represented by risp::RispValue.

Decimal integers are read as risp::RispValue::Int(...)

assert_eq!(risp::read("123").unwrap()[0], risp::RispValue::Int(123));

Numbers with decimal points are considered risp::RispValue::Float(...)

assert_eq!(risp::read("3.14").unwrap()[0], risp::RispValue::Float(3.14));

Symbols use a variety of characters, but they cannot start with a digit.

assert_eq!(risp::read("foo-BAR123?<>!@$%^&*").unwrap()[0],
           risp::RispValue::Symbol(String::from("foo-BAR123?<>!@$%^&*")));

String literals begin and end with quotation marks (").

assert_eq!(risp::read("\"foo\"").unwrap()[0],
           risp::RispValue::String(String::from("foo")));

Escaping is handled for ", , n, r, and t.

assert_eq!(risp::read("\"\\\"\\\\\\n\\r\\t\"").unwrap()[0], risp::RispValue::String(String::from("\"\\\n\r\t")));

An empty list is parsed as Nil

assert_eq!(risp::read("()").unwrap()[0], risp::RispValue::Nil);

A list of elemens is a linked list of Cons

assert_eq!(risp::read("(1 2)").unwrap()[0],
           risp::RispValue::Cons(
               std::sync::Arc::new(risp::RispValue::Int(1)),
               std::sync::Arc::new(risp::RispValue::Cons(std::sync::Arc::new(risp::RispValue::Int(2)),
                                        std::sync::Arc::new(risp::RispValue::Nil)))));

Brackes enclose a Cons pair

assert_eq!(risp::read("[1 2]").unwrap()[0],
           risp::RispValue::Cons(std::sync::Arc::new(risp::RispValue::Int(1)),
                                 std::sync::Arc::new(risp::RispValue::Int(2))));

A single quote (') prefixing an s-expressions constructs a quote special form.

use risp::*;
use std::sync::Arc;
assert_eq!(read("'(a b c)").unwrap()[0],
           read("(quote (a b c))").unwrap()[0]);

Whitespace includes spaces, tabs and line-ends (LF and CR).

risp::read("(fn foo (x y)\n\r\t(+ x y))").unwrap();

risp::read() returns a vector containing all s-expressions in the input

assert_eq!(risp::read("1 two \"three\"").unwrap(), vec![risp::RispValue::Int(1),
                                                              risp::RispValue::Symbol(String::from("two")),
                                                              risp::RispValue::String(String::from("three"))]);